+++ .circleci/config.yml
@@ -240,7 +240,7 @@ jobs:
           key: typescript-declaration-files-{{ .Branch }}-{{ .Revision }}
           paths:
             # packages with generated declaration files
-            - packages/mui-base/build
+            - packages/react/build
   test_types_next:
     <<: *default-job
     resource_class: 'medium+'
@@ -272,7 +272,7 @@ jobs:
       - run:
           name: Diff declaration files
           command: |
-            git add -f packages/mui-base/build || echo '/base declarations do not exist'
+            git add -f packages/react/build || echo 'Base UI declarations do not exist'
             pnpm -r build:types
             git --no-pager diff
       - run:
