+++ .eslintignore
@@ -2,7 +2,7 @@
 /coverage
 /docs/export
 /docs/pages/playground/
-/packages/mui-base/**/*.min.*
+/packages/react/**/*.min.*
 /packages/mui-utils/macros/__fixtures__/
 # Ignore fixtures
 /test/bundling/fixtures/**/*.fixture.js
