+++ .github/CODEOWNERS
@@ -1,4 +1,4 @@
 * @atomiks @michaldudak
 /docs/data/ @atomiks @colmtuite @michaldudak
-/packages/mui-base/ @atomiks @colmtuite @michaldudak
+/packages/react/ @atomiks @colmtuite @michaldudak
 /scripts/ @michaldudak
