+++ docs/src/app/experiments/slider-marks.tsx
@@ -1,7 +1,7 @@
 'use client';
 import * as React from 'react';
 import { Slider } from '@base-ui-components/react/Slider';
-import { useSliderRootContext } from '../../../../packages/mui-base/src/Slider/Root/SliderRootContext';
+import { useSliderRootContext } from '../../../../packages/react/src/Slider/Root/SliderRootContext';
 
 const STOPS = [
   {
