+++ docs/tsconfig.json
@@ -35,7 +35,7 @@
   "exclude": ["node_modules"],
   "references": [
     {
-      "path": "../packages/mui-base/tsconfig.build.json"
+      "path": "../packages/react/tsconfig.build.json"
     }
   ]
 }
