+++ tsconfig.base.json
@@ -15,8 +15,8 @@
     "allowJs": true,
     "skipLibCheck": true,
     "paths": {
-      "@base-ui-components/react": ["./packages/mui-base/src"],
-      "@base-ui-components/react/*": ["./packages/mui-base/src/*"],
+      "@base-ui-components/react": ["./packages/react/src"],
+      "@base-ui-components/react/*": ["./packages/react/src/*"],
       "@mui-internal/api-docs-builder": ["./node_modules/@mui/monorepo/packages/api-docs-builder"],
       "@mui-internal/api-docs-builder/*": [
         "./node_modules/@mui/monorepo/packages/api-docs-builder/*"
