+++ tsconfig.json
@@ -7,7 +7,7 @@
       "path": "docs/tsconfig.json"
     },
     {
-      "path": "packages/mui-base/tsconfig.json"
+      "path": "packages/react/tsconfig.json"
     }
   ]
 }
